+++ test/ring/swagger/swagger2_test.clj
@@ -2,6 +2,7 @@
   (:require [schema.core :as s]
             [ring.swagger.swagger2 :refer :all]
             [ring.swagger.validator :as v]
+            [flatland.ordered.map :as om]
             [ring.util.http-status :as status]
             [midje.sweet :refer :all])
   (:import [java.util Date UUID]
@@ -291,3 +292,10 @@
                  :paths {"/pet"   {:post {:tags ["pet"]}}
                          "/store" {:post {:tags ["store"]}}}}]
     (validate swagger) => nil))
+
+(fact "route order"
+  (let [->path (fn [x] (str "/" x))
+        paths (reduce (fn [acc x] (assoc acc (->path x) {:get {}})) (om/ordered-map) (range 100))
+        swagger {:paths paths}
+        spec (swagger-json swagger)]
+    (-> spec :paths keys) => (map ->path(range 100))))